arxiv/requirements.txt

# Python dependencies for arXiv frontend

# arXiv API client
arxiv==2.1.3

# MySQL database connector
pymysql==1.1.0

# Flask web framework
Flask==3.0.0

# Python dotenv for configuration
python-dotenv==1.0.0

# HTTP requests
requests

# Production WSGI server
gunicorn==21.2.0
